+++ bsp/win32/port/luat_gpio_win32.c
@@ -29,7 +29,7 @@ int luat_gpio_setup(luat_gpio_t* gpio) {
     return 0;
 }
 
-int luat_gpio_set(int pin, int level) {
+int luat_gpio_set(int pin, uint8_t level) {
     if (pin < 0 || pin >= LUAT_WIN32_GPIO_COUNT) {
         return -1;
     }
